+++ src/Calculators/Dominion/MilitaryCalculator.php
@@ -242,8 +242,8 @@ class MilitaryCalculator
         $multiplier = 0;
 
         // Values (percentages)
-        $opPerGryphonNest = 1.75;
-        $gryphonNestMaxOp = 35;
+        $opPerGryphonNest = 1.6;
+        $gryphonNestMaxOp = 32;
 
         if ($dominion->calc !== null && !isset($dominion->calc['invasion'])) {
             if (isset($dominion->calc['gryphon_nest_percent'])) {
@@ -500,8 +500,8 @@ class MilitaryCalculator
         $multiplier = 0;
 
         // Values (percentages)
-        $dpPerGuardTower = 1.75;
-        $guardTowerMaxDp = 35;
+        $dpPerGuardTower = 1.6;
+        $guardTowerMaxDp = 32;
 
         if ($dominion->calc !== null && !isset($dominion->calc['invasion'])) {
             if (isset($dominion->calc['guard_tower_percent'])) {
@@ -600,8 +600,8 @@ class MilitaryCalculator
     public function getTempleReduction(Dominion $dominion): float
     {
         // Values (percentages)
-        $dpReductionPerTemple = 1.5;
-        $templeMaxDpReduction = 25;
+        $dpReductionPerTemple = 1.35;
+        $templeMaxDpReduction = 27;
         $dpMultiplierReduction = 0;
 
         if ($dominion->calc !== null && !isset($dominion->calc['invasion'])) {